Update to increase realism of environment
The update includes a revision of the environment to now include
- a geological model that is based on 2 random walks: Volumentric joint count and UCS
- potential for random failure of cutters due to blocky rockmass behavior

The blocky rockmass model was implemented following the models from Delisio & Zhao (2014) - "A new model for TBM performance prediction in blocky rock conditions". Based on this blockyness is implemented as well as a revision of the penetration which governs the wear of the cutters.